lizans with a lizard face can now gain a snake(-like) tongue.
One reason it to make it play even more nicely with ascending as a fully TFed dragon. On the other hand it was a pita, when you want to be a
lizan with dragon wings and a snake-like tongue using Snake Oil, Reptilium and Drakes Heart.
Lowered the dragonScore()-requirement to be considered being a dragon morph from >= 6 to >= 5. This is final now.
Added the method gainSnakeTongue() to the MutationsHelper-class since I'm using that code snippet twice now.