Finished converting players rearBody into a Bodypart container class
This is mainly as a POC to show, how converting body parts into a container class basically works.

Savefile data is converted and not backwards compatiable so, when testing this you should always backup your saves or errors may occur upon loading a savegame or save to an empty slot during testing.